qrp/backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
whitenoise>=6.5.0
Pillow>=10.0.0
gunicorn==21.2.0
django-redis>=5.3.0
